* prelim fix nvarchar(max) on PriceListDetai
* Add decimal precision and scale

+++ src/DMSpro.OMS.MdmService.Application.Contracts/PriceListDetails/PriceListDetailDto.cs
@@ -6,8 +6,8 @@ namespace DMSpro.OMS.MdmService.PriceListDetails
 {
     public class PriceListDetailDto : FullAuditedEntityDto<Guid>, IHasConcurrencyStamp
     {
-        public int Price { get; set; }
-        public int? BasedOnPrice { get; set; }
+        public decimal Price { get; set; }
+        public decimal? BasedOnPrice { get; set; }
         public string Description { get; set; }
         public Guid PriceListId { get; set; }
         public Guid UOMId { get; set; }
